feat(bitflips): Addressed issues when the central DN and mode were separated.
If the mode is outside the medstd boundary, we should not use the maxima
boundaries from it. Instead pick the next highest prominence that is
within the medstd boundary to set the maxima boundaries. When these two
central tendencies are within the medstd boundary, don't select any
minima between them.